moved AbstractSimulationConfig to de.uka.ipd.sdq.simulation.core
