+++ bundles/de.uka.ipd.sdq.simucomframework.simulationdock/src/de/uka/ipd/sdq/simucomframework/simulationdock/SimulationProgressReportingObserver.java
@@ -6,8 +6,8 @@ import java.util.Map;
 import org.osgi.service.event.Event;
 import org.osgi.service.event.EventAdmin;
 
-import de.uka.ipd.sdq.simulation.AbstractSimulationConfig;
 import de.uka.ipd.sdq.simulation.IStatusObserver;
+import de.uka.ipd.sdq.simulation.core.AbstractSimulationConfig;
 
 public class SimulationProgressReportingObserver implements IStatusObserver {
     private int lastPercent = 0;
