+++ bundles/de.uka.ipd.sdq.simulation.core/META-INF/MANIFEST.MF
@@ -3,6 +3,12 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Palladio Simulation Core
 Bundle-SymbolicName: de.uka.ipd.sdq.simulation.core
 Bundle-Version: 5.2.1.qualifier
+Export-Package: de.uka.ipd.sdq.simulation.core
+Require-Bundle: org.apache.log4j;bundle-version="1.2.19",
+ org.palladiosimulator.analyzer.workflow.core;bundle-version="5.2.1",
+ org.palladiosimulator.recorderframework;bundle-version="5.2.1",
+ de.uka.ipd.sdq.probfunction.math;bundle-version="5.2.1",
+ de.uka.ipd.sdq.simulation.abstractsimengine;bundle-version="5.2.1"
 Bundle-Vendor: edu.kit.ipd.sdq
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
