+++ bundles/de.uka.ipd.sdq.simulation.core/src/de/uka/ipd/sdq/simulation/core/ISimulationListener.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.simulation;
+package de.uka.ipd.sdq.simulation.core;
 
 public interface ISimulationListener {
