+++ bundles/de.uka.ipd.sdq.simulation/META-INF/MANIFEST.MF
@@ -13,7 +13,8 @@ Require-Bundle: org.eclipse.ui,
  de.uka.ipd.sdq.probfunction.math;bundle-version="2.0.0",
  de.uka.ipd.sdq.simulation.abstractsimengine;bundle-version="1.0.0",
  org.palladiosimulator.analyzer.workflow.core;bundle-version="1.0.0",
- de.uka.ipd.sdq.simucomframework.simucomstatus;bundle-version="1.0.0"
+ de.uka.ipd.sdq.simucomframework.simucomstatus;bundle-version="1.0.0",
+ de.uka.ipd.sdq.simulation.core;bundle-version="5.2.1";visibility:=reexport
 Bundle-ActivationPolicy: lazy
 Export-Package: de.uka.ipd.sdq.simulation,
  de.uka.ipd.sdq.simulation.preferences
