+++ bundles/de.uka.ipd.sdq.simulation/src/de/uka/ipd/sdq/simulation/ISimulationControl.java
@@ -1,6 +1,7 @@
 package de.uka.ipd.sdq.simulation;
 
 import de.uka.ipd.sdq.simucomframework.simucomstatus.SimuComStatus;
+import de.uka.ipd.sdq.simulation.core.AbstractSimulationConfig;
 
 /**
  * Control interface for simulation frameworks. It defines methods, which are responsible for
